# backend/api/docs/swagger.yaml
basePath: /
definitions:
  models.CommentPost:
    properties:
      value:
        type: string
    required:
    - value
    type: object
  models.CreateOrUpdatePost:
    properties:
      message:
        minLength: 5
        type: string
      selectedFile:
        type: string
      title:
        type: string
    required:
    - message
    - title
    type: object
...
  models.PostModel:
    properties:
      _id:
        type: string
      comments:
        items:
          type: string
        type: array
      createdAt:
        type: string
      creator:
        type: string
      likes:
        items:
          type: string
        type: array
      message:
        minLength: 5
        type: string
      name:
        type: string
      selectedFile:
        type: string
      title:
        type: string
    required:
    - message
    - title
    type: object
...
host: localhost:5005
info:
  contact: {}
  description: This is Swagger docs for rest api goland fiber
  title: Fiber Goland Mongo Grpc Websocet etc..
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: get all post
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: user id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: get all post
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: create new post
      parameters:
      - description: post create details
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrUpdatePost'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: delete post by post id need to provided auth token for post creator
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: delete post by id
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: get post details
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: get post data
      tags:
      - Post
    patch:
      consumes:
      - application/json
      description: update post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: update post details
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CreateOrUpdatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: update post
      tags:
      - Posts
  /posts/{id}/commentPost:
    post:
      consumes:
      - application/json
      description: comment post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: comment value
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.CommentPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: comment post
      tags:
      - Posts
  /posts/{id}/likePost:
    patch:
      consumes:
      - application/json
      description: like post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: like or unlike a post by it's id
      tags:
      - Posts
  /posts/search:
    get:
      consumes:
      - application/json
      description: get post and users matching the search query
      parameters:
      - description: Search query
        in: query
        name: searchQuery
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: get post by search query
      tags:
      - Posts
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"